+++ src/prelude.rs
@@ -1,3 +1,4 @@
+pub use crate::aligner::alignment::table::{Table, TraceFlags, TracebackTable};
 pub use crate::aligner::alignment::{Alignment, SSWResult, Traceback};
 pub use crate::aligner::{Aligner, AlignerBuilder};
 pub use crate::error::{Error, Result};